Completed database, completed get all spots, get all spots owned by the current user, create spot, add an image to a spot based on the spot id and edit a spot

+++ backend/db/migrations/20240105004331-create-spot.js
@@ -29,10 +29,10 @@ module.exports = {
         type: Sequelize.STRING
       },
       lat: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL(9,7)
       },
       lng: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL(10,7)
       },
       name: {
         type: Sequelize.STRING
